Create news pages for action list and new action: - Renamed current custom actions pages
