+++ src/domain/entities/User.ts
@@ -1,4 +1,3 @@
-import _ from "lodash";
 import { NamedRef } from "../../domain/entities/Ref";
 
 export interface User {
