+++ src/webapp/pages/landing/LandingPage.tsx
@@ -2,11 +2,9 @@ import React, { useMemo } from "react";
 import { useHistory } from "react-router-dom";
 import i18n from "../../../locales";
 import { Card, CardGrid } from "../../components/card-grid/CardGrid";
-import { useAppContext } from "../../contexts/app-context";
 
 export const LandingPage: React.FC = () => {
     const history = useHistory();
-    const { compositionRoot } = useAppContext();
 
     const cards: Card[] = useMemo(
         () => [
@@ -17,7 +15,7 @@ export const LandingPage: React.FC = () => {
                         name: i18n.t("Actions"),
                         description: i18n.t("Actions are a way to trigger events in your application"),
                         listAction: () => history.push("/actions"),
-                        addAction: () => compositionRoot.actions.exampleAction().run(console.debug, console.error),
+                        addAction: () => history.push("/actions/new"),
                     },
                     {
                         name: i18n.t("Browse xMART"),
@@ -27,7 +25,7 @@ export const LandingPage: React.FC = () => {
                 ],
             },
         ],
-        [history, compositionRoot]
+        [history]
     );
 
     return <CardGrid cards={cards} />;
