+++ src/webapp/pages/list-custom-actions/ListCustomActionsPage.tsx
@@ -1,6 +1,9 @@
-import { ObjectsTable, TableAction, TableColumn, useSnackbar } from "@eyeseetea/d2-ui-components";
+import { ObjectsTable, TableAction, TableColumn, useLoading, useSnackbar } from "@eyeseetea/d2-ui-components";
+import { Sync } from "@material-ui/icons";
+import _ from "lodash";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useHistory } from "react-router-dom";
+import { Future } from "../../../domain/entities/Future";
 import { SyncAction } from "../../../domain/entities/SyncAction";
 import { SyncResult } from "../../../domain/entities/SyncResult";
 import i18n from "../../../locales";
@@ -10,6 +13,7 @@ import { useAppContext } from "../../contexts/app-context";
 export const ListActionsPage: React.FC = () => {
     const { compositionRoot } = useAppContext();
     const snackbar = useSnackbar();
+    const loading = useLoading();
     const history = useHistory();
 
     const [rows, setRows] = useState<SyncAction[]>([]);
@@ -23,10 +27,37 @@ export const ListActionsPage: React.FC = () => {
         []
     );
 
-    const actions: TableAction<SyncAction>[] = useMemo(() => [], []);
+    const actions: TableAction<SyncAction>[] = useMemo(
+        () => [
+            {
+                name: "run",
+                text: i18n.t("Run"),
+                icon: <Sync />,
+                multiple: true,
+                onClick: (ids: string[]) => {
+                    loading.show(true, i18n.t("Running actions..."));
+                    const futures = _.compact(ids.map(id => rows.find(item => item.id === id))).map(({ execute }) =>
+                        execute()
+                    );
+
+                    Future.parallel(futures, { maxConcurrency: 1 }).run(
+                        results => {
+                            loading.reset();
+                            setResults(results);
+                        },
+                        error => {
+                            loading.reset();
+                            snackbar.error(error);
+                        }
+                    );
+                },
+            },
+        ],
+        [snackbar, loading, rows]
+    );
 
     const goToCreateAction = useCallback(() => {
-        history.push("/actions/new");
+        history.push("/custom-actions/new");
     }, [history]);
 
     useEffect(() => {
